fix: some refactoring and improvements
