+++ index.js
@@ -1,7 +1,9 @@
+const router = require('express').Router();
+
 const analyticsRouter = require('./routers/analytics');
 const tasksRouter = require('./routers/tasks');
 
-module.exports = {
-  analyticsRouter,
-  tasksRouter,
-};
+router.get('/analytics', analyticsRouter);
+router.get('/tasks', tasksRouter);
+
+module.exports = router;
