+++ routers/analytics.js
@@ -1,7 +1,6 @@
 const router = require('express').Router();
-const { getAnalytics, createTask } = require('../controllers/analytics');
+const { getAnalytics } = require('../controllers/analytics');
 
 router.get('/', getAnalytics);
-router.post('/', createTask);
 
 module.exports = router;
