+++ routers/tasks.js
@@ -1,6 +1,7 @@
 const router = require('express').Router();
-const getTask = require('../controllers/tasks');
+const { getTask, createTask } = require('../controllers/tasks');
 
 router.get('/', getTask);
+router.post('/', createTask);
 
 module.exports = router;
